place analytics action in same file
https://coveord.atlassian.net/browse/KIT-676
