+++ packages/headless/src/features/result-preview/result-preview-actions.ts
@@ -8,7 +8,13 @@ import {
   buildResultPreviewRequest,
   StateNeededByHtmlEndpoint,
 } from './result-preview-request-builder';
-import {logDocumentQuickview} from './result-preview-analytics-actions';
+import {
+  AnalyticsType,
+  documentIdentifier,
+  makeAnalyticsAction,
+  partialDocumentInformation,
+  validateResultPayload,
+} from '../analytics/analytics-utils';
 
 interface FetchResultContentResponse {
   content: string;
@@ -51,3 +57,20 @@ export const fetchResultContent = createAsyncThunk<
     };
   }
 );
+
+const logDocumentQuickview = (result: Result) => {
+  return buildDocumentQuickviewThunk(result)();
+};
+
+const buildDocumentQuickviewThunk = (result: Result) => {
+  return makeAnalyticsAction(
+    'analytics/resultPreview/open',
+    AnalyticsType.Click,
+    (client, state) => {
+      validateResultPayload(result);
+      const info = partialDocumentInformation(result, state);
+      const id = documentIdentifier(result);
+      return client.logDocumentQuickview(info, id);
+    }
+  );
+};
